+++ presentation/src/main/java/com/mashup/gabbangzip/sharedalbum/presentation/ui/main/mypage/component/GroupItemVersion.kt
@@ -1,4 +1,4 @@
-package com.mashup.gabbangzip.sharedalbum.presentation.ui.main.mypage
+package com.mashup.gabbangzip.sharedalbum.presentation.ui.main.mypage.component
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.mashup.gabbangzip.sharedalbum.presentation.theme.Gray60
@@ -14,29 +13,26 @@ import com.mashup.gabbangzip.sharedalbum.presentation.theme.Gray80
 import com.mashup.gabbangzip.sharedalbum.presentation.theme.PicTypography
 
 @Composable
-fun UserContainer(
+fun GroupItemVersion(
     modifier: Modifier = Modifier,
-    userName: String,
-    loginWay: String,
+    titleText: String,
+    versionName: String,
 ) {
-    Column(
-        modifier = modifier,
-        horizontalAlignment = Alignment.Start,
-    ) {
+    Column(modifier = modifier) {
         Text(
             modifier = Modifier
                 .fillMaxWidth()
                 .wrapContentHeight(),
-            text = userName,
-            style = PicTypography.headBold20,
+            text = titleText,
+            style = PicTypography.bodyMedium16,
             color = Gray80,
         )
         Text(
             modifier = Modifier
                 .fillMaxWidth()
                 .wrapContentHeight()
-                .padding(top = 4.dp),
-            text = loginWay,
+                .padding(top = 9.dp),
+            text = versionName,
             style = PicTypography.captionNormal12,
             color = Gray60,
         )
