+++ presentation/src/main/java/com/mashup/gabbangzip/sharedalbum/presentation/ui/main/mypage/component/GroupTitle.kt
@@ -1,4 +1,4 @@
-package com.mashup.gabbangzip.sharedalbum.presentation.ui.main.mypage
+package com.mashup.gabbangzip.sharedalbum.presentation.ui.main.mypage.component
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
