+++ presentation/src/main/java/com/mashup/gabbangzip/sharedalbum/presentation/ui/main/mypage/component/UserContainer.kt
@@ -1,4 +1,4 @@
-package com.mashup.gabbangzip.sharedalbum.presentation.ui.main.mypage
+package com.mashup.gabbangzip.sharedalbum.presentation.ui.main.mypage.component
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -6,6 +6,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.mashup.gabbangzip.sharedalbum.presentation.theme.Gray60
@@ -13,26 +14,29 @@ import com.mashup.gabbangzip.sharedalbum.presentation.theme.Gray80
 import com.mashup.gabbangzip.sharedalbum.presentation.theme.PicTypography
 
 @Composable
-fun GroupItemVersion(
+fun UserContainer(
     modifier: Modifier = Modifier,
-    titleText: String,
-    versionName: String,
+    userName: String,
+    loginWay: String,
 ) {
-    Column(modifier = modifier) {
+    Column(
+        modifier = modifier,
+        horizontalAlignment = Alignment.Start,
+    ) {
         Text(
             modifier = Modifier
                 .fillMaxWidth()
                 .wrapContentHeight(),
-            text = titleText,
-            style = PicTypography.bodyMedium16,
+            text = userName,
+            style = PicTypography.headBold20,
             color = Gray80,
         )
         Text(
             modifier = Modifier
                 .fillMaxWidth()
                 .wrapContentHeight()
-                .padding(top = 9.dp),
-            text = versionName,
+                .padding(top = 4.dp),
+            text = loginWay,
             style = PicTypography.captionNormal12,
             color = Gray60,
         )
